Introduce python replicas of ibd_main.cc for testing
