+++ pylib/load.py
@@ -1,4 +1,5 @@
 import ROOT
+ROOT.PyConfig.IgnoreCommandLineOptions = True
 ROOT.gSystem.Load('build/libIbdSel.so')
 
 # ROOT.TH1.AddDirectory( False )
