+++ selector/ClusterAlg.cc
@@ -1,65 +1,3 @@
-#pragma once
+#include "ClusterAlg.hh"
 
-// TODO: Also save singles here
-
-#include "RingBuf.hh"
-#include "Util.hh"
-
-#include "EventReader.hh"
-
-using Status = Algorithm::Status;
-
-class ClusterAlg : public SimpleAlg<EventReader> {
-  static constexpr float GAPSIZE_US = 1000;
-
-public:
-  ClusterAlg(int detector) : detector(detector) {}
-
-  Status consume(const EventReader::Data& e) override;
-
-  bool ready() const { return foundCluster; };
-
-  const std::vector<EventReader::Data>& getCluster() const
-  {
-    return events;
-  }
-
-  const short detector;
-
-private:
-  std::vector<EventReader::Data> events;
-  EventReader::Data pendingEvent;
-  bool foundCluster = false;
-};
-
-inline Status ClusterAlg::consume(const EventReader::Data& e)
-{
-  if (foundCluster) {                 // Did we find a cluster on prev exec?
-    foundCluster = false;
-    events.clear();
-    events.push_back(pendingEvent);
-  }
-
-  if (e.detector != detector)
-    return Status::Continue;
-
-  if (e.energy < 0.7)
-    return Status::Continue;
-
-  if (events.size() == 0) {     // Very first event
-    events.push_back(e);
-    return Status::Continue;
-  }
-
-  const Time t_this = e.time();
-  const Time t_last = events.back().time();
-
-  if (t_this.diff_us(t_last) > GAPSIZE_US) {
-    pendingEvent = e;
-    foundCluster = true;
-  } else {
-    events.push_back(e);
-  }
-
-  return Status::Continue;
-}
+template ClusterAlg& makeAlg<ClusterAlg>(Pipeline&, int);
